squashed a silly warning
git-svn-id: http://otx.osxninja.com/builds/trunk@189 5968870d-8a1b-0410-8ef2-9208d525af3e

+++ source/Processors/X86Processor.m
@@ -1959,7 +1959,7 @@
 //	UInt8	opcode	= inCode[0];
 //	UInt8	opcode2	= inCode[1];
 
-	return IS_JUMP(inCode[0], inCode[1]);
+	return IS_JUMP((UInt8)inCode[0], (UInt8)inCode[1]);
 }
 
 //	gatherFuncInfos
